feat: backend security

+++ core/services/srv-tag/srv-tag-impl/src/main/java/esthesis/services/tag/impl/service/TagService.java
@@ -1,9 +1,13 @@
 package esthesis.services.tag.impl.service;
 
+import esthesis.common.AppConstants.Security.Category;
+import esthesis.common.AppConstants.Security.Operation;
+import esthesis.common.exception.QSecurityException;
 import esthesis.service.common.BaseService;
 import esthesis.service.common.paging.Page;
 import esthesis.service.common.paging.Pageable;
 import esthesis.service.common.validation.CVExceptionContainer;
+import esthesis.service.security.resource.SecurityResource;
 import esthesis.service.tag.entity.TagEntity;
 import esthesis.services.tag.impl.repository.TagRepository;
 import esthesis.util.kafka.notifications.common.KafkaNotificationsConstants.Action;
@@ -16,6 +20,7 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.microprofile.jwt.JsonWebToken;
+import org.eclipse.microprofile.rest.client.inject.RestClient;
 
 @Slf4j
 @ApplicationScoped
@@ -27,21 +32,43 @@ public class TagService extends BaseService<TagEntity> {
   @Inject
   TagRepository tagRepository;
 
+  @Inject
+  @RestClient
+  SecurityResource securityResource;
+
   @Override
   public Page<TagEntity> find(Pageable pageable) {
     log.debug("Finding all tags with '{}'.", pageable);
-    return super.find(pageable);
+
+    if (!securityResource.isPermitted(Category.TAG, Operation.READ)) {
+      throw new QSecurityException("You are not allowed to view tags.");
+    } else {
+      return super.find(pageable);
+    }
   }
 
   @Override
   public Page<TagEntity> find(Pageable pageable, boolean partialMatch) {
     log.debug("Finding all tags with partial match with '{}'.", pageable);
-    return super.find(pageable, partialMatch);
+    if (!securityResource.isPermitted(Category.TAG, Operation.READ)) {
+      throw new QSecurityException("You are not allowed to view tags.");
+    } else {
+      return super.find(pageable, partialMatch);
+    }
   }
 
   @Override
   public TagEntity save(TagEntity tagEntity) {
     log.debug("Saving tag '{}'.", tagEntity);
+
+    // Security check.
+    if (tagEntity.getId() == null && !securityResource.isPermitted(Category.TAG,
+        Operation.CREATE)) {
+      throw new QSecurityException("You are not allowed to create tags.");
+    } else if (!securityResource.isPermitted(Category.TAG, Operation.WRITE)) {
+      throw new QSecurityException("You are not allowed to update tags.");
+    }
+
     // Ensure no other tag has the same name.
     TagEntity existingTagEntity = findFirstByColumn("name", tagEntity.getName());
     if (existingTagEntity != null && (tagEntity.getId() == null || !existingTagEntity.getId()
@@ -58,18 +85,30 @@ public class TagService extends BaseService<TagEntity> {
       idParamOrder = 0, payload = "Tag ID")
   public boolean deleteById(String id) {
     log.debug("Deleting tag with id '{}'.", id);
-    return super.deleteById(id);
+    if (!securityResource.isPermitted(Category.TAG, Operation.DELETE, id)) {
+      throw new QSecurityException("You are not allowed to delete this tags.");
+    } else {
+      return super.deleteById(id);
+    }
   }
 
   public List<TagEntity> findByName(String name, boolean partialMatch) {
-    return findByName(Collections.singletonList(name), partialMatch);
+    if (!securityResource.isPermitted(Category.TAG, Operation.READ)) {
+      throw new QSecurityException("You are not allowed to view tags.");
+    } else {
+      return findByName(Collections.singletonList(name), partialMatch);
+    }
   }
 
   public List<TagEntity> findByName(List<String> names, boolean partialMatch) {
-    if (partialMatch) {
-      return tagRepository.findByNamePartial(names);
+    if (!securityResource.isPermitted(Category.TAG, Operation.READ)) {
+      throw new QSecurityException("You are not allowed to view tags.");
     } else {
-      return tagRepository.findByName(names);
+      if (partialMatch) {
+        return tagRepository.findByNamePartial(names);
+      } else {
+        return tagRepository.findByName(names);
+      }
     }
   }
 }
